Reflect suggestion
